+++ ledger-protocol/src/main/kotlin/com/ledgerpasswords/companion/ledger/transport/FakeLedgerTransport.kt
@@ -8,12 +8,12 @@ class FakeLedgerTransport(
     private val appName: String = "Passwords",
     private val appVersion: String = "1.3.1",
     private val storageSize: Int = LedgerPasswordsLimits.DEFAULT_STORAGE_SIZE,
-    initialMetadatas: ByteArray = ByteArray(storageSize) { 0x00 },
+    initialMetadatas: ByteArray = ByteArray(storageSize) { 0x00.toByte() },
 ) : LedgerTransport {
     private var metadatas: ByteArray = initialMetadatas.copyOf(storageSize)
     private var dumpOffset: Int = 0
     private var loadOffset: Int = 0
-    private val loadBuffer: ByteArray = ByteArray(storageSize) { 0x00 }
+    private val loadBuffer: ByteArray = ByteArray(storageSize) { 0x00.toByte() }
 
     override suspend fun exchange(cla: Int, ins: Int, p1: Int, p2: Int, data: ByteArray): ApduResponse = when {
         cla == ApduConstants.CLA_SDK && ins == ApduConstants.INS_GET_APP_INFO -> appInfo()
@@ -27,7 +27,7 @@ class FakeLedgerTransport(
     private fun appInfo(): ApduResponse {
         val name = appName.toByteArray(Charsets.US_ASCII)
         val version = appVersion.toByteArray(Charsets.US_ASCII)
-        val payload = byteArrayOf(0x01, name.size.toByte()) + name + byteArrayOf(version.size.toByte()) + version
+        val payload = byteArrayOf(0x01.toByte(), name.size.toByte()) + name + byteArrayOf(version.size.toByte()) + version
         return ApduResponse(payload, StatusWords.OK)
     }
 
@@ -40,8 +40,8 @@ class FakeLedgerTransport(
             ((size ushr 16) and 0xFF).toByte(),
             ((size ushr 8) and 0xFF).toByte(),
             (size and 0xFF).toByte(),
-            0x00,
-            0x00,
+            0x00.toByte(),
+            0x00.toByte(),
         )
         return ApduResponse(payload, StatusWords.OK)
     }
@@ -69,7 +69,7 @@ class FakeLedgerTransport(
         loadOffset += data.size
         if (p1 == ApduConstants.LAST_CHUNK || loadOffset >= storageSize) {
             metadatas = loadBuffer.copyOf(storageSize)
-            loadBuffer.fill(0x00)
+            loadBuffer.fill(0x00.toByte())
             loadOffset = 0
         }
         return ApduResponse(byteArrayOf(), StatusWords.OK)
